Вынос response в модели и переименование тестов
