+++ internal/models/responses.go
@@ -1,0 +1,12 @@
+package models
+
+type SuccessResponse struct {
+	Status int
+	Data   interface{}
+}
+
+type ErrResponse struct {
+	Status int
+	Msg    string
+	MsgRus string
+}
